% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotators.R
\name{word_annotator}
\alias{easy_parse_annotator}
\alias{parse_annotator}
\alias{word_annotator}
\title{Annotators}
\usage{
word_annotator()

parse_annotator()

easy_parse_annotator(n = 2)
}
\arguments{
\item{n}{The number of generations to go back (see \code{\link[base]{parent.frame}}).}
}
\value{
Returns an annotator for entities or words.

\code{easy_parse_annotator} - Assigns a parse annotator to \code{.parse_ann}
in the global nvironment if not found.
}
\description{
\code{word_annotator} - A wrapper for \code{\link[openNLP]{Maxent_Word_Token_Annotator}}.

\code{parse_annotator} - A wrapper for \code{\link[openNLP]{Parse_Annotator}}.

\code{easy_parse_annotator} - A wrapper for \code{\link[openNLP]{Parse_Annotator}}
that checks for \code{.parse_ann} in the global environment.  If \code{.parse_ann}
is not found a copy is assigned to the global environment for future sourcing.
This is because the parse annotator often leds to a Java out of memory error
if multiple instances are assigned.
}
\seealso{
\code{\link[openNLP]{Parse_Annotator}},
\code{\link[openNLP]{Maxent_Word_Token_Annotator}}
}

