% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_square_brace.R
\name{as_square_brace}
\alias{as_square_brace}
\alias{as_square_brace_latex}
\title{Conver parser Output to Square Braces}
\usage{
as_square_brace(x, ...)

as_square_brace_latex(x, ...)
}
\arguments{
\item{x}{A parse vector from \code{parser}.}

\item{\ldots}{ignored.}
}
\value{
\code{as_square_brace} - returns a vector appropriate for use in a
many online parse tree generators.

\code{as_square_brace_latex} - returns a \code{\link[base]{cat}}
vector appropriate for use in a LaTeX document.
}
\description{
Many tree parsers (e.g., online, LaTeX, etc.) require square braces in the
parse.  These functions prepare the parse for use in other plotting venues.
}
\note{
When using \code{as_square_brace_latex}, the user will need to add the
following LaTeX packages to the document: qtree, color, xcolor.  Additionally,
the user will need to add the code chunk, \code{\\newenvironment\{dummy\}\{\}\{\}},
to the preamble.
}
\examples{
\dontrun{
## pass `as_square_brace` to an online parse tree:
## http://yohasebe.com/rsyntaxtree
if(!exists('parse_ann')) {
    parse_ann <- parse_annotator()
}
x <- parser("Really, I like chocolate because it is good.", parse_ann)
xsquared <- as_square_brace(x)

## get tree from web
if (!require("pacman")) install.packages("pacman"); library(pacman)
p_load_gh("ropensci/RSelenium")
p_load(XML, RSelenium)

# download Selenium Server, if not already present
checkForServer(); Sys.sleep(2)

# Open a remote browser
remDr <- remoteDriver(
    remoteServerAddr = "localhost",
    port = 4444,
    browserName = "firefox"
)

## MIGHT NEED:
## RSelenium::startServer()

remDr$open()

## Now we pass the url from above and the page is opened
remDr$navigate("http://yohasebe.com/rsyntaxtree/")

webElem <- remDr$findElement("css", "textarea[id='data']")
webElem$clearElement()
webElem$sendKeysToElement(list(xsquared)); Sys.sleep(.5)


webElem2 <- remDr$findElement("css", "button[id='draw_png']")
webElem2$clickElement()

## LaTeX Version
## copy .Rmd from parser package
treeloc <- system.file("extra_files/testtree.Rmd", package = "parsent")
file.copy(treeloc, getwd())

## look at .Rmd
file.edit("testtree.Rmd")

## render .Rmd file
p_load(rmarkdown)
rmarkdown::render("testtree.Rmd")
}
}

