% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tree.R
\name{as_tree}
\alias{as_tree}
\title{Convert a Parsed String to Tree}
\usage{
as_tree(x, ...)
}
\arguments{
\item{x}{A vector from \code{parser}.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of trees from \code{\link[NLP]{Tree_parse}}.
}
\description{
Convert a parsed string to a tree structure.  A wrapper for
\code{\link[NLP]{Tree_parse}}.
}
\examples{
\dontrun{
txt <- c(
    "Really, I like chocolate because it is good. It smells great.",
    "Robots are rather evil and most are devoid of decency.",
    "He is my friend.",
    "Clifford the big red dog ate my lunch.",
    "Professor Johns can not teach",
    "",
    NA
)

if(!exists('parse_ann')) {
    parse_ann <- parse_annotator()
}
x <- parser(txt, parse_ann)
as_tree(x)
}
}
\seealso{
\code{\link[NLP]{Tree_parse}}
}
\keyword{parse}
\keyword{tree}

