% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_leaves.R
\name{get_leaves}
\alias{get_leaves}
\title{Extract Tokens from a Phrase}
\usage{
get_leaves(x, regex = "@tokens")
}
\arguments{
\item{x}{A list/vetor of phrases}

\item{regex}{A regular expression to extract tokens.  Default extracts tokens:
\code{"(?<=\\s)[A-Za-z'-]+(?=\\))"}.  Use \code{"(?<=\\s)[A-Za-z'-]+(?=\\))"}
to extract words.  Use short hand \code{regex = "@words"} to extract words or
\code{regex = "@tokens"} to extract tokens.}
}
\value{
Returns a list of vectors of extracted tokens.
}
\description{
Extract the tokens from a phrase.
}
\examples{
\dontrun{
txt <- c(
    "Really, I like chocolate because it is good. It smells great.",
    "Robots are rather evil and most are devoid of decency.",
    "He is my friend.",
    "Clifford the big red dog ate my lunch.",
    "Professor Johns can not teach",
    "",
    NA
)

if(!exists('parse_ann')) {
    parse_ann <- parse_annotator()
}
(x <- parser(txt, parse_ann))

get_leaves(get_phrase_type_regex(x, "NP"))

## As a dplyr chain
library(dplyr)
x \%>\%
    get_phrase_type_regex("NP") \%>\%
    get_leaves()

## Just words (in this case no difference)
x \%>\%
    get_phrase_type_regex("NP") \%>\%
    get_leaves("@words")
}
}
\keyword{leaves}
\keyword{tokens}
\keyword{words}

