% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_phrases.R
\name{get_phrases}
\alias{get_phrases}
\title{Extract the Phrases from Parsed Sentence}
\usage{
get_phrases(x)
}
\arguments{
\item{x}{A parsed character string or list (see \code{parser}).}
}
\value{
Returns a list of character vectors of highest level phrases.
}
\description{
Extract highest level phrases from a sentence (see
\url{http://www.surdeanu.info/mihai/teaching/ista555-spring15/readings/PennTreebankConstituents.html}
for more on phrase types).
}
\examples{
\dontrun{
txt <- c(
    "Really, I like chocolate because it is good. It smells great.",
    "Robots are rather evil and most are devoid of decency.",
    "He is my friend.",
    "Clifford the big red dog ate my lunch.",
    "Professor Johns can not teach",
    "",
    NA
)

parse_ann <- parse_annotator()
(x <- parser(txt, parse_ann))
get_phrases(x)

lapply(get_phrases(x), get_leaves)
}
}
\keyword{phrases}

