% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{plot.parsed_character}
\alias{plot.parsed_character}
\title{Plots a plot.parsed_character Object}
\usage{
\method{plot}{parsed_character}(x, vertex.color = NA,
  vertex.frame.color = NA, vertex.label.font = 2, vertex.label.cex = 1,
  edge.width = 1.5, edge.color = "black", edge.arrow.size = 0,
  leaf.color = "chartreuse4", phrase.marker.color = "blue4", title = NULL,
  cex.title = 0.9, asp = 0.5, ...)
}
\arguments{
\item{x}{A \code{parsed_character} object (see \code{\link[parsent]{parser}}.}

\item{vertex.color}{The vertex color (see \code{?igraph::igraph_options}).}

\item{vertex.frame.color}{The vertex frame color (see \code{?igraph::igraph_options}).}

\item{vertex.label.font}{The vertex label font (see \code{?igraph::igraph_options}).}

\item{vertex.label.cex}{The vertex label scaled relative to the default (see
\code{?igraph::igraph_options}).}

\item{edge.width}{The edge width (see \code{?igraph::igraph_options}).}

\item{edge.color}{The edge color (see \code{?igraph::igraph_options}).}

\item{edge.arrow.size}{The edge arrow size (see \code{?igraph::igraph_options}).}

\item{leaf.color}{The color of the leaves (tokens).}

\item{phrase.marker.color}{The color of the non-terminal grammar categories
(e.g., NP, VP, D, etc.).}

\item{title}{The main title of the graph.}

\item{cex.title}{The size of the title relative to the default.}

\item{asp}{The y/x aspect ratio.}

\item{\ldots}{Other arguments passed to \code{\link[igraph]{plot.igraph}}.}
}
\description{
Plots a plot.parsed_character object
}
\references{
\url{http://stackoverflow.com/a/33536291/1000343}
}
\author{
StackOverflow's \href{http://stackoverflow.com/users/2415684/thetime}{TheTime} and Tyler Rinker <tyler.rinker@gmail.com>.
}
