% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{print.parsed_character}
\alias{print.parsed_character}
\title{Prints a parsed_character Object}
\usage{
\method{print}{parsed_character}(x, ...)
}
\arguments{
\item{x}{A parsed_character Object.}

\item{\ldots}{ignored.}
}
\description{
Prints a parsed_character object
}

