% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotators.R
\name{word_annotator}
\alias{parse_annotator}
\alias{word_annotator}
\title{Annotators}
\usage{
word_annotator()

parse_annotator()
}
\value{
Returns an annotator for entities or words.
}
\description{
A wrapper for \code{\link[openNLP]{Parse_Annotator}} and
\code{\link[openNLP]{Maxent_Word_Token_Annotator}}.
}
\seealso{
\code{\link[openNLP]{Parse_Annotator}},
\code{\link[openNLP]{Maxent_Word_Token_Annotator}}
}

