% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsent-package.R
\docType{package}
\name{parsent}
\alias{parsent}
\alias{package-parsent}
\alias{parsent-package}
\title{Sentence Parsing Tools}
\description{
Tools to parse sentences.  A wrapper for \pkg{NLP}/\pkg{openNLP} that
simplifies and extends the user experience.
}
